Added implementation of the RunLengthEncode operator.
RunLengthEncode transforms a sequence of values into a sequence of KeyValuePair<T,int> representing the occurrence of repeated runs of the same value. An overload is provided that accepts an IEqualityComparer allowing the caller to supply the equality semantics for elements in the sequence. RunLengthEncode performs its transformation in a deferred, streaming fashion - and can operate on sequences of arbitrary (or infinite) length.