Added implementation of the SegmentOperator.
The Segment operator divides a sequence into multiple sequences by using a caller-supplied segmentation detector. Individual segments are buffered, but the overall sequence is returned using a streaming implementation. Overloads of segment exist that supply the segmentation function with the index of the element as well as the previous element in the sequence, to allow positional or differential division of a sequence.