Added implementation of the Lag operator.
Lag projects a sequence by combining each element with a prior element at a given offset. For elements prior to th lag offset, the lagged value is considered to be default(T). An overload exists that allows the caller to supply an alternate default value. Lag behaves in a deferred and streaming manner.