Added implementation of the Rank and RankBy operators.
Rank returns a sequences of positive numbers representing the rank of items in the source sequence, with the lowest values corresponding to the "highest" value in the sequence. RankBy works similarly by allows ranking to be based on a key determined by a caller supplied function. Overload also exist that accept an IComparer<T> allowing the comparison semantics to be defined by the caller.